+++ client/front/lib/calculate_margin.dart
