+++ client/front/lib/fixerrand/fix_errand_widget/fixdue.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:front/fixerrand/fixerrand.dart';
+import 'package:front/fixerrand/fix_errand.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 class FixDue extends StatefulWidget{
