+++ client/front/lib/fixerrand/fix_errand_widget/fixminimap.dart
