+++ client/front/lib/showerrand/show_errand.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 import 'dart:convert';
 import 'dart:developer';
-import 'package:front/showerrand/showerrandwidget/show_errand_widget.dart';
+import 'package:front/showerrand/show_errand_widget/show_errand_widget.dart';
 import 'package:http/http.dart' as http;
 
 import 'package:flutter/cupertino.dart';
