+++ client/front/lib/showerrand/show_errand_widget/stamp/stamp.dart
